fixed problem: when deleting foreign item no access was given to url and shortLink entries + refactoring
